ENH: BrainMask module logic improved.
The K-Means and Bayesian segmentation works fine for all the tissues desired.
However, the selection for one tissue class still need to be fixed.
It having troubles with the copy the segmented maps, in order to create a single class map.
XML file was improved to add the Segmentation Parameters tab.