NEW: BrainExtractionTool added.
Copy from ROBEX brain extraction tool.
The module exits in Slicer when it is finished.
Maybe it has problems from memory release, in the final parte of the algorithm.

TODO: The ref_vols and dat folders need to be generalized in the start method.
See the header in BrainExtractionTool.cxx file.