BUG: AAD applied on DWI images now is functioning.
The general image processing, applying the AAD filter on DWI images, are working in this version.
However, the output image, seen in the DTI reconstruction, reveals some misleading values.
The FA maps appears as much higher than the original image. Maybe the rescale factor is doing something in the pixel values.