BUG: Code stability regarding conductance parameter
The Conductance parameters was corrected regarding the data type, which were set to float precision.
Now the AAD filtering is able to handle a larger range of input datasets.
In addition, the DiffusionEdgeOptimizationImageCalculator class was also added to the CLI code in order to
set the conductance parameter automatically, using 3 optimization functions.